# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KnowBase API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/documents/sources": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_documents_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceOut"
                  },
                  "type": "array",
                  "title": "Response List Sources Api Documents Sources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/documents/upload": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_api_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/documents/ingest-url": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Ingest Url",
        "operationId": "ingest_url_api_documents_ingest_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/URLIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/documents/source/{source_name}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Source",
        "operationId": "delete_source_api_documents_source__source_name__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/clear": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Clear Kb",
        "operationId": "clear_kb_api_documents_clear_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/knowledge-base/stats": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Stats",
        "operationId": "stats_api_knowledge_base_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KBStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/knowledge-base/chunks": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Chunks",
        "operationId": "chunks_api_knowledge_base_chunks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Source"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chunks Api Knowledge Base Chunks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge-base/sources": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "List Source Names",
        "operationId": "list_source_names_api_knowledge_base_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response List Source Names Api Knowledge Base Sources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/knowledge-base/config": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Kb Config",
        "operationId": "kb_config_api_knowledge_base_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KBConfig"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/knowledge-base/hotspots": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Hotspots",
        "operationId": "hotspots_api_knowledge_base_hotspots_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HotspotEntry"
                  },
                  "type": "array",
                  "title": "Response Hotspots Api Knowledge Base Hotspots Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_file_api_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_documents_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HotspotEntry": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "hits": {
            "type": "integer",
            "title": "Hits"
          },
          "content_preview": {
            "type": "string",
            "title": "Content Preview"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "source",
          "hits",
          "content_preview"
        ],
        "title": "HotspotEntry",
        "description": "Hot chunk entry returned by GET /api/knowledge-base/hotspots."
      },
      "IngestResponse": {
        "properties": {
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "total_docs": {
            "type": "integer",
            "title": "Total Docs"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "chunk_count",
          "total_docs",
          "message"
        ],
        "title": "IngestResponse"
      },
      "KBConfig": {
        "properties": {
          "chunk_size": {
            "type": "integer",
            "title": "Chunk Size"
          },
          "chunk_overlap": {
            "type": "integer",
            "title": "Chunk Overlap"
          }
        },
        "type": "object",
        "required": [
          "chunk_size",
          "chunk_overlap"
        ],
        "title": "KBConfig",
        "description": "Knowledge base configuration returned by GET /api/knowledge-base/config."
      },
      "KBStats": {
        "properties": {
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count"
          },
          "total_chars": {
            "type": "integer",
            "title": "Total Chars"
          }
        },
        "type": "object",
        "required": [
          "chunk_count",
          "source_count",
          "total_chars"
        ],
        "title": "KBStats"
      },
...
      "SourceOut": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "source",
          "count"
        ],
        "title": "SourceOut",
        "description": "Document source summary returned by GET /api/documents/sources."
      },
      "URLIngestRequest": {
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1,
            "title": "Url",
            "description": "Public URL to fetch and ingest"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "URLIngestRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}